subteam limits

+++ server/api/subteam/views.py
@@ -6,10 +6,18 @@ from .serializers import SubTeamSerialiser
 
 @api_view(['POST'])
 def create_subteam(request):
-    serialiser = SubTeamSerialiser(data=request.data)
-    if serialiser.is_valid():
-        serialiser.save()
-        return Response(serialiser.data)
+    if (request.user.is_authenticated is False):
+        return Response("User not authenticated", status=401)
+    else:
+        if (request.user.team_admin is False):
+            return Response("User is not authorised to create a subteam", status=403)
+        else:
+            serialiser = SubTeamSerialiser(data=request.data)
+            if serialiser.is_valid():
+                serialiser.save()
+                return Response(serialiser.data)
+            else:
+                return Response(serialiser.errors, status=400)
 
 
 @api_view(['GET'])
@@ -21,15 +29,29 @@ def get_subteams(request, team_id):
 
 @api_view(['PUT'])
 def update_subteam(request, subteam_id):
-    subteam = SubTeam.objects.get(subteam_id=subteam_id)
-    serializer = SubTeamSerialiser(instance=subteam, data=request.data)
-    if serializer.is_valid():
-        serializer.save()
-        return Response(serializer.data)
+    if (request.user.is_authenticated is False):
+        return Response("User not authenticated", status=401)
+    else:
+        if (request.user.team_admin is False):
+            return Response("User is not authorised to update a subteam", status=403)
+        else:
+            subteam = SubTeam.objects.get(subteam_id=subteam_id)
+            serializer = SubTeamSerialiser(instance=subteam, data=request.data)
+            if serializer.is_valid():
+                serializer.save()
+                return Response(serializer.data)
+            else:
+                return Response(serializer.errors, status=400)
 
 
 @api_view(['DELETE'])
 def delete_subteam(request, subteam_id):
-    subteam = SubTeam.objects.get(subteam_id=subteam_id)
-    subteam.delete()
-    return Response("SubTeeam successfully deleted")
+    if (request.user.is_authenticated is False):
+        return Response("User not authenticated", status=401)
+    else:
+        if (request.user.team_admin is False):
+            return Response("User is not authorised to delete a subteam", status=403)
+        else:
+            subteam = SubTeam.objects.get(subteam_id=subteam_id)
+            subteam.delete()
+            return Response("SubTeeam successfully deleted")
